fix(hardening): 深度闭环跨平台CI、科学正确性、凭据隔离、硬件执行与文档一致性
- CI 跨平台与基础设施升级 (P1-01):
  - tests/test_harness_neutral.py: 修正 Windows 下测试子进程源码非 ASCII 解码失败
  - .github/workflows/qa.yml: macos-13 迁移至 macos-15-intel，追加 Python 3.14 矩阵支持
  - .github/workflows/live-contract.yml: 统一为 actions/checkout 和 setup-python 完整 SHA 固定
- MCP 科学统计正确性与学术真诚度 (P1-02, P1-03):
  - scripts/mcp_server.py: 修正调用不存在函数 check_p_value_match 为 check_p_match
  - scripts/mcp_server.py: 解包 p_from_t receipt 提取真实浮点数 p 值
  - scripts/mcp_server.py: 严格要求 6 项参数全量提供才计算 Cohen's d，拒绝伪造假数据
  - scripts/mcp_server.py: inputSchema 放宽 df 为 number，支持 Welch 非整数自由度
- Hermes 子进程生命周期与质询参数修复 (P1-07, P1-08):
  - hermes_cli_adapter.py: 增加超时进程 terminate -> kill 回收清理，杜绝孤儿后台进程
  - hermes_cli_adapter.py: 独立模式 challenge 严格使用 task_path 作为 prompt 参数
- 环境变量隔离去跨 Provider 污染 (P1-05):
  - contracts.py: 按 declared provider 建立显式白名单，阻断跨厂商凭据交叉泄漏
- 硬件执行层与科学浮点容差 (P1-09, P1-10):
  - workload_profiles.py: 接通 torch_xpu -> 'xpu' 真实执行分支
  - workload_profiles.py: 区分 float32/float64 数值门禁容差 (float32 为 1e-5~2e-4)，避免误杀加速器
- 学术引用监控三态表达 (P1-11):
  - literature-watch: OpenAlex 缺失反向引用时如实上报不可观测，杜绝折叠为 0 引用假阴性
- 契约与文档全面对齐 (P2, P3):
  - contracts.py / schemas: PanelSpec 统一为嵌套 budget 结构，review-finding 增加 P3 严重级别
  - README (8 份多语言) 全面刷新为 Harness-Neutral 架构定位与 470 项测试，纠正韩文무선 효과为무작위 효과

+++ scripts/scfabric/admission.py
@@ -100,7 +100,7 @@ def run_admission(workload, scale, dtype="float64", gain_threshold=1.5,
             })
             continue
 
-        parity_ok = wp.check_parity(workload, scale, ref_result, cand_result)
+        parity_ok = wp.check_parity(workload, scale, ref_result, cand_result, dtype=dtype)
         if not parity_ok:
             candidates.append({
                 "backend": name,
